+++ test/fixtures/cookbooks/firewall-test/attributes/default.rb
@@ -1,0 +1,1 @@
+default['firewall_test_iptables'] = false