+++ test/fixtures/cookbooks/firewall-test/recipes/debian_iptables.rb
@@ -1,14 +1,3 @@
-#########
-# firewall
-#########
-Chef::Platform.set platform: :ubuntu, resource: :firewall, provider: Chef::Provider::FirewallIptables
-
-#########
-# firewall_rule
-#########
-Chef::Platform.set platform: :ubuntu, resource: :firewall_rule, provider: Chef::Provider::FirewallRuleIptables
-
-log 'save iptables rules' do
-  level :debug
-  notifies :save, 'firewall[default]', :delayed
-end
+Chef::Platform.set platform: :ubuntu, resource: :firewall, provider: Chef::Provider::FirewallIptablesUbuntu
+Chef::Platform.set platform: :ubuntu, resource: :firewall_rule, provider: Chef::Provider::FirewallRuleIptablesUbuntu
+node.override['firewall_test_iptables'] = true