+++ test/integration/default/serverspec/iptables_spec.rb
@@ -3,10 +3,9 @@ require_relative 'spec_helper'
 
 expected_rules = [
   # we included the .*-j so that we don't bother testing comments
-  %r{-A INPUT -p tcp -m tcp -m multiport --dports 7788 .*-j ACCEPT},
-  %r{-A INPUT -p tcp -m tcp -m state --state RELATED,ESTABLISHED .*-j ACCEPT},
+  %r{-A INPUT -m state --state RELATED,ESTABLISHED .*-j ACCEPT},
   %r{-A INPUT -p tcp -m tcp -m multiport --dports 22 .*-j ACCEPT},
-  %r{-A INPUT -p tcp -m tcp -m multiport --dports 2222,2200 .*-j ACCEPT},
+  %r{-A INPUT -p tcp -m tcp -m multiport --dports 2200,2222 .*-j ACCEPT},
   %r{-A INPUT -p tcp -m tcp -m multiport --dports 1234 .*-j DROP},
   %r{-A INPUT -p tcp -m tcp -m multiport --dports 1235 .*-j REJECT --reject-with icmp-port-unreachable},
   %r{-A INPUT -p tcp -m tcp -m multiport --dports 1236 .*-j DROP},
@@ -15,10 +14,9 @@ expected_rules = [
 ]
 
 expected_ipv6_rules = [
-  %r{-A INPUT( -s ::/0 -d ::/0)? -p tcp -m tcp -m multiport --dports 7788 .*-j ACCEPT},
-  %r{-A INPUT( -s ::/0 -d ::/0)? -p tcp -m tcp -m state --state RELATED,ESTABLISHED .*-j ACCEPT},
+  %r{-A INPUT( -s ::/0 -d ::/0)? -m state --state RELATED,ESTABLISHED .*-j ACCEPT},
   %r{-A INPUT( -s ::/0 -d ::/0)? -p tcp -m tcp -m multiport --dports 22 .*-j ACCEPT},
-  %r{-A INPUT( -s ::/0 -d ::/0)? -p tcp -m tcp -m multiport --dports 2222,2200 .*-j ACCEPT},
+  %r{-A INPUT( -s ::/0 -d ::/0)? -p tcp -m tcp -m multiport --dports 2200,2222 .*-j ACCEPT},
   %r{-A INPUT( -s ::/0 -d ::/0)? -p tcp -m tcp -m multiport --dports 1234 .*-j DROP},
   %r{-A INPUT( -s ::/0 -d ::/0)? -p tcp -m tcp -m multiport --dports 1235 .*-j REJECT --reject-with icmp6-port-unreachable},
   %r{-A INPUT( -s ::/0 -d ::/0)? -p tcp -m tcp -m multiport --dports 1236 .*-j DROP},
@@ -37,7 +35,7 @@ describe command('iptables-save'), :if => iptables? do
   duplicate_rule1 = '-A INPUT -p tcp -m tcp -m multiport --dports 1111 -m comment --comment "same comment" -j ACCEPT'
   its(:stdout) { should count_occurences(duplicate_rule1, 1) }
 
-  duplicate_rule2 = '-A INPUT -p tcp -m tcp -m multiport --dports 5432,5431 -m comment --comment "same comment" -j ACCEPT'
+  duplicate_rule2 = '-A INPUT -p tcp -m tcp -m multiport --dports 5431,5432 -m comment --comment "same comment" -j ACCEPT'
   its(:stdout) { should count_occurences(duplicate_rule2, 1) }
 end